FIXED BUG in feature RegionSampling, leading to measured kin. energy of drift is always zero.
git-svn-id: https://scm.projects.hlrs.de/authscm/jenkinsls1/svn/ls1/MarDyn/branches/mheinenNEMD/mainbranch20170119@6123 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/NEMD/RegionSampling.cpp
@@ -1251,7 +1251,7 @@ void SampleRegion::CalcGlobalValuesProfiles(DomainDecompBase* domainDecomp, Doma
 					_d2EkinTransComp[offsetSum+s] += d2EkinTrans;
 					// Ekin drift
 					unsigned long N = _nNumMoleculesGlobal[offsetN+s];
-					unsigned long dInvN = 1. / ( (double)(N) );
+					double dInvN = 1. / ( (double)(N) );
 					double vd = _dVelocityGlobal[offsetComp+s];
 					double d2EkinDrift = vd*vd * _vecMass.at(cid) * dInvN;
 					_d2EkinDriftComp[offsetComp+s] = d2EkinDrift;
